+++ Projects/Android/jni/OpenJK/codeJK2/game/g_navigator.h
@@ -246,7 +246,7 @@ protected:
 	char	GetChar( fileHandle_t file );
 	int		GetInt( fileHandle_t file );
 	float	GetFloat( fileHandle_t file );
-	long	GetLong( fileHandle_t file );
+	uint32_t	GetLong( fileHandle_t file );
 
 	//void	ConnectNodes( void );
 	void	SetEdgeCost( int ID1, int ID2, int cost );
